Add a feature to create orders with item values

+++ src/module/order/application/service/order.service.ts
@@ -6,17 +6,39 @@ import {
 import { Order, OrderStatus } from '../../domain/order.entity';
 import { UserService } from '../../../../module/user/application/service/user.service';
 import { RoleEnum } from '../../../../module/user/domain/user.entity';
-import { CannotUpdateOrderException } from '../errors/CannotUpdateOrder';
+import {
+  CannotAccessOrderException,
+  CannotUpdateOrderException,
+} from '../errors/CannotUpdateOrder';
+import { Item } from 'src/module/item/domain/item.entity';
+import { ProductService } from 'src/module/product/application/service/product.service';
+import { ItemService } from 'src/module/item/application/service/item.service';
+
 @Injectable()
 export class OrderService {
   constructor(
     @Inject(ORDER_REPOSITORY) private orderRepository: IOrderRepository,
     @Inject(UserService) private userService: UserService,
+    @Inject(ProductService) private productService: ProductService,
+    @Inject(ItemService) private itemService: ItemService,
   ) {}
 
   async delete(userId: number, orderId: number): Promise<void> {
-    await this.findById(userId, orderId);
-    return await this.orderRepository.delete(orderId);
+    const user = await this.userService.findUserById(userId);
+    if (user.role === RoleEnum.admin) {
+      const orderToDelete = await this.findById(userId, orderId);
+      if (!orderToDelete) {
+        throw new NotFoundException();
+      }
+      await this.itemService.deleteRelation(orderToDelete.items);
+      await this.orderRepository.delete(orderId);
+    }
+    const order = user.orders.find((order) => order.id === orderId);
+    if (!order) {
+      throw new CannotAccessOrderException();
+    }
+    await this.itemService.deleteRelation(order.items);
+    await this.orderRepository.delete(orderId);
   }
 
   async findAll(userId: number): Promise<Order[]> {
@@ -45,8 +67,17 @@ export class OrderService {
 
   async create(userId: number, order: Order): Promise<Order> {
     const user = await this.userService.findUserById(userId);
-    order.user = user;
-    return await this.orderRepository.create(order);
+    const newOrder = new Order();
+
+    const savedOrder = await this.orderRepository.create(newOrder);
+    savedOrder.items = await this.saveItems(order.items, savedOrder);
+    const total = savedOrder.items.reduce((acc, item) => {
+      return acc + item.subTotal;
+    }, 0);
+    savedOrder.total = total;
+    savedOrder.user = user;
+    const createOrder = await this.orderRepository.save(savedOrder);
+    return createOrder;
   }
 
   async update(userId: number, orderId: number, updatedOrder: Order) {
@@ -56,11 +87,27 @@ export class OrderService {
     }
     const order = user.orders.find((order) => order.id === orderId);
     if (!order) {
-      throw new NotFoundException();
+      throw new CannotAccessOrderException();
     }
     if (order.status !== OrderStatus.pending) {
       throw new CannotUpdateOrderException();
     }
     return await this.orderRepository.update(orderId, updatedOrder);
+  }
+
+  async saveItems(items: Item[], savedOrder: Order): Promise<Item[]> {
+    const savedItems = items.map(async (item: Item) => {
+      const newItem = new Item();
+      const product = await this.productService.getOne(item.product.id);
+      const subtotal = product.price * item.quantity;
+      newItem.product = product;
+      newItem.order = savedOrder;
+      newItem.subTotal = subtotal;
+      newItem.quantity = item.quantity;
+      const savedItem = await this.itemService.save(newItem);
+      return savedItem;
+    });
+    const result = await Promise.all(savedItems);
+    return result;
   }
 }
